+++ libs/domain/lib/event-listener.ts
@@ -1,7 +1,7 @@
 import { Event } from './event';
 
 export abstract class EventListener {
-  abstract getEventTypePrefixes(): string[]
+  abstract getStreamNamePrefixes(): string[]
 
   async on(event: Event): Promise<void> {
     const methodName = `on${event.type}`;