+++ libs/domain/lib/query.ts
@@ -1,4 +1,5 @@
 export interface Query {
   readonly type: string
   readonly by: string
+  readonly timestamp: string
 }