+++ libs/domain/test-domain/test.event-listener.ts
@@ -1,13 +1,8 @@
 import { EventListener } from '../lib/event-listener';
 
 export class TestEventListener extends EventListener {
-  // eslint-disable-next-line class-methods-use-this
-  getEventTypePrefixes(): string[] {
-    return [];
-  }
+  getStreamNamePrefixes(): string[] { return []; }
 
   // eslint-disable-next-line class-methods-use-this
-  async onTestEvent(): Promise<void> {
-    return Promise.resolve();
-  }
+  async onTestEvent(): Promise<void> { return Promise.resolve(); }
 }