+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.39
+
+- add `timestamp` to `Query` and `Command`
+- add `getStreamNamePrefx` to `Aggregate`
+- add `getStreamNamePrefixes` to `EventListener`
+- fix `pl-file-form` #66
+
+---
+
 ## 0.0.38
 
 - fix `InMemoryFeatureService` adapter inheritance